Implemented sweet alert

+++ src/app/components/tutorial-details/tutorial-details.component.ts
@@ -2,6 +2,7 @@ import { Component, Input, OnInit } from '@angular/core';
 import { TutorialService } from 'src/app/services/tutorial.service';
 import { ActivatedRoute, Router } from '@angular/router';
 import { Tutorial } from 'src/app/models/tutorial.model';
+import Swal from 'sweetalert2';
 
 @Component({
   selector: 'app-tutorial-details',
@@ -66,25 +67,60 @@ export class TutorialDetailsComponent implements OnInit {
   updateTutorial(): void {
     this.message = '';
 
-    this.tutorialService.update(this.currentTutorial.id, this.currentTutorial)
-      .subscribe({
-        next: (res) => {
-          console.log(res);
-          this.message = res.message ? res.message : 'This tutorial was updated successfully!';
-        },
-        error: (e) => console.error(e)
+    Swal.fire({
+      title: 'Are you sure you want to update this Tutorial?',
+      text: "Data will be updated!",
+      icon: 'warning',
+      showCancelButton: true,
+      confirmButtonColor: '#3085d6',
+      cancelButtonColor: '#d33',
+      confirmButtonText: 'Yes, update it!'
+    }).then((result) => {
+      if (result.isConfirmed) {
+        this.tutorialService.update(this.currentTutorial.id, this.currentTutorial)
+        .subscribe({
+          next: (res) => {
+            console.log(res);
+            this.message = res.message ? res.message : 'This tutorial was updated successfully!';
+            Swal.fire(
+              'Updted!',
+              'Tutorial has been updted.',
+              'success'
+            )
+          },
+          error: (e) => console.error(e)
       });
+      }
+    });
   }
 
   deleteTutorial(): void {
-    this.tutorialService.delete(this.currentTutorial.id)
-      .subscribe({
-        next: (res) => {
-          console.log(res);
-          this.router.navigate(['/tutorials']);
-        },
-        error: (e) => console.error(e)
-      });
+    Swal.fire({
+      title: 'Are you sure you want to delete this Tutorial?',
+      text: "You won't be able to revert this!",
+      icon: 'warning',
+      showCancelButton: true,
+      confirmButtonColor: '#3085d6',
+      cancelButtonColor: '#d33',
+      confirmButtonText: 'Yes, delete it!'
+    }).then((result) => {
+      if (result.isConfirmed) {
+        this.tutorialService.delete(this.currentTutorial.id)
+          .subscribe({
+            next: (res) => {
+              console.log(res);
+              Swal.fire(
+                'Deleted!',
+                'Your Tutorial has been deleted.',
+                'success'
+              )
+              this.router.navigate(['/tutorials']);
+            },
+            error: (e) => console.error(e)
+          });
+      }
+    })
+    
   }
 
 }
